ai_module/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langgraph
python-dotenv
langchain-aws
